debugged the errors and db successfully taking all the inputs

+++ src/utils/cloudinary.js
@@ -1,13 +1,16 @@
 import {v2 as cloudinary} from 'cloudinary';
 import { response } from 'express';
 import fs from "fs"
-cloudinary.config({
-    Cloud_Name: process.env.Cloud_Name,
-    API_Key: process.env.API_Key,
-    API_Secret: process.env.API_Secret 
-})
+
 
 export const uploadOnCloudinary=async (localFilePath) => {
+    
+    
+    cloudinary.config({
+    cloud_name: process.env.cloud_name,
+    api_key: process.env.api_key,
+    api_secret: process.env.api_secret 
+})
     try{
 
         if(!localFilePath)return null;
@@ -16,10 +19,11 @@ export const uploadOnCloudinary=async (localFilePath) => {
                 resource_type:"auto"
             })
             console.log("file is uploaded on cloudinary",response.url);
+            fs.unlinkSync(localFilePath); 
             return response;
 }catch(error){
     console.log("cannot upload file",error);
-    fs.unlinkSync(localFilePath);
+   if(fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
     return null;
 }
 }
